Remove dependency of lib freetype
Now we use Qt for generating the font bitmaps. With less dependencies,
we are one step closer to working on multiple platforms; this also makes
it easier to install the plugin. Finally, it could be possible in the
near future to customize the font just by setting its name in the
configuration file.