Solve compilation issues on Mac OS
Unfortunately, the architecture adopted for GDB ImageWatch means that UI calls
are attempted at a secondary thread, which is not supported on MacOS (and is
prone to bugs on Linux). We should probably refactor the code, having two
separate processes (one for GDB and another for the UI), using IPC between
them.