feat: gdb support class/struct members
lldb already auto complete for classes. This idea is to have something
similar for gdb.

As an example, if 'this' is derived from a base class, the auto
completion will show 'this.member' instead of showing only 'member'.

If the scope have a class or a struct that is observable the behavior is
not changed.

If the scope have a class with observable member, the member will be
auto completed as class.member.

In order to test this behavior the testbench is slightly modified to add
a new class with observable members.